Increment run_id subroot TTL and treat expiration as non-retriable.
Summary:
This changes the TTL for run_id subroot from 30s to 2 hours --> the value was always intended to be in this order of magnitude, and I only had it set to 30s for iterating locally.

Additionally, as was discovered during runs on AWS, some calls into `torch.distributed` appear to be blocking and do not release the GIL -- this means that the ephemeral key lease renewal thread in python does not run, and can cause expiration. Generally this is not an issue because the job has catastrophically failed anyway, but the resulting logs (missing key `version_counter`) would be misleading. This diff fixes that with a non-retryable exception.

Reviewed By: kiukchung

Differential Revision: D18738369

fbshipit-source-id: 93e6d7e5335af405912b6f0d0e6874d4a993d1f8