expose run_id as TORCHELASTIC_RUN_ID env var for workers (#103)
Summary:
Pull Request resolved: https://github.com/pytorch/elastic/pull/103

run_id is a unique id for the run instance of the job. It is useful for the workers to have access to this information for various purposes. For instance it serves a a prefix for a directory that local rank 0 creates that would not collide with other jobs that are stacked on the same node.

Reviewed By: theweiho

Differential Revision: D22104173

fbshipit-source-id: 0c81b0f2e936813cb1cc7692721f2cc6ebbf1222