Move torchelastic.distributed.app to torch.distributed.fb.app (#106)
Summary:
Pull Request resolved: https://github.com/pytorch/elastic/pull/106

D7881729 breaks torchelastic's circle ci tests in oss because `Future` moved from `torch.rpc` to `torch.futures` which is not available in torch 1.5.0 which is what torchelastic depends on in oss (1.5.0 is the latest stable release at the time of this writing).

Since we were planning to move `torchelastic.distributed.app` to distributed PT anyways, do this now (incubated under fb dir).

Note the `base_dir` on the module is still made to be `torch.distributed` so that when we drop the `fb` directory we don't have to go codemod all the usages.

Reviewed By: mrshenli

Differential Revision: D22148675

fbshipit-source-id: f8529d8c2ecbe8653203650358f09ef2c54f49f8