1/n Make fb init_metrics to accept custom context
Summary:
This change adds custom context to the internal to fb init_metric method. It will allow to record metrics not only when the job is executed in flow, but also on any other platform.

The alternative way is to utilize the environment variables, but I think this way it makes things cleaner to the users.

Usage:

  from torchelastic import metrics
  metrics.initialize_metrics("metric_group")

Reviewed By: kiukchung

Differential Revision: D22077102

fbshipit-source-id: ab65756d45a0efaaee450b21e3e15139408ddd94