Multiprocess library
Summary:
Add initial functionality of the popen multiprocessing library. The library uses popen to start child processes and error reporter to record errors.

Two termination behaviors are supported: GROUP and SINGLE.
SINGLE - each process in a group is treated separately, and single failure does not cause group failure.
GROUP - each process is a part of group, meaning that if a single process dies, the whole group will be terminated.

TODO: in future diff we need to add ability to raise a root cause error out of all the errors returned by child processes.

Reviewed By: taohe

Differential Revision: D24066717

fbshipit-source-id: e7dd227d430f0136eb804b3a12d334ee1407c83e